+++ tools/cli/installers/lib/ide/shared/task-tool-command-generator.js
@@ -176,7 +176,6 @@ class TaskToolCommandGenerator {
 
     return `---
 description: '${description.replaceAll("'", "''")}'
-disable-model-invocation: true
 ---
 
 # ${item.displayName || item.name}